Original kernel 2.4.37.5

+++ release/src/linux/linux/include/asm-sh/hdreg.h
@@ -7,6 +7,7 @@
 #ifndef __ASM_SH_HDREG_H
 #define __ASM_SH_HDREG_H
 
-typedef unsigned int ide_ioreg_t;
+//typedef unsigned int ide_ioreg_t;
+typedef unsigned long ide_ioreg_t;
 
 #endif /* __ASM_SH_HDREG_H */
